Move using directive out of header
Putting a using directive in a header affects the namespaces of any file
that includes that header. Instead, we can let the user decide whether or
not they want all of std in their namespace by moving our using directives
to our source files. This also avoids potential namespace conflicts in user
code.